Refactor log rotation tests from `error_spam_test.go` into `functional_test.go` and `error_spam_helpers.go`.
